requirements.txt
fastapi
uvicorn
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
sqlalchemy
psycopg2-binary
alembic 